refactor: 컴포넌트 폴더 구조 및 스타일 파일 분리
