+++ src/components/home/FriendSelectBox/styles.ts
@@ -1,8 +1,7 @@
+import { css } from "@emotion/react";
 import type { Theme } from "@emotion/react";
-import { css, useTheme } from "@emotion/react";
-import { FiPlus } from "react-icons/fi";
 
-const boxStyle = (theme: Theme) => css`
+export const boxStyle = (theme: Theme) => css`
   display: flex;
   align-items: center;
   background-color: ${theme.colors.semantic.kakaoYellow};
@@ -11,7 +10,7 @@ const boxStyle = (theme: Theme) => css`
   margin-bottom: ${theme.spacing.spacing6};
 `;
 
-const plusIconStyle = (theme: Theme) => css`
+export const plusIconStyle = (theme: Theme) => css`
   background-color: ${theme.colors.gray.gray00};
   color: ${theme.colors.gray.gray1000};
   border-radius: 50%;
@@ -24,19 +23,7 @@ const plusIconStyle = (theme: Theme) => css`
   margin-right: ${theme.spacing.spacing3};
 `;
 
-const textStyle = (theme: Theme) => css`
+export const textStyle = (theme: Theme) => css`
   font-size: ${theme.typography.body1Bold.fontSize};
   font-weight: ${theme.typography.body1Bold.fontWeight};
 `;
-
-export default function FriendSelectBox() {
-  const theme = useTheme();
-  return (
-    <div css={boxStyle(theme)}>
-      <div css={plusIconStyle(theme)}>
-        <FiPlus />
-      </div>
-      <span css={textStyle(theme)}>선물할 친구를 선택해 주세요.</span>
-    </div>
-  );
-}
