+++ src/components/home/RankingSection/components/RankingFilterTabs/index.tsx
@@ -1,33 +1,10 @@
-import type { Theme } from "@emotion/react";
-import { css, useTheme } from "@emotion/react";
+import { useTheme } from "@emotion/react";
 import { useSearchParams } from "react-router-dom";
 import { useEffect } from "react";
+import { tabsWrapper, tabItemStyle } from "./styles";
 
 const FILTER_LABELS = ["전체", "여성", "남성", "청소년이"] as const;
 const FILTER_OPTIONS = FILTER_LABELS;
-
-const tabsWrapper = (theme: Theme) => css`
-  display: flex;
-  justify-content: space-around;
-  padding: 0 ${theme.spacing.spacing4} ${theme.spacing.spacing3};
-  font-size: ${theme.fontSizes.label1Regular};
-`;
-
-const tabItem = (theme: Theme) => css`
-  cursor: pointer;
-  padding: 6px 8px;
-  border-radius: 8px;
-  font-weight: 500;
-
-  &:hover {
-    background-color: ${theme.colors.gray.gray200};
-  }
-`;
-
-const tabItemStyle = (theme: Theme, isSelected: boolean) => css`
-  ${tabItem(theme)};
-  background-color: ${isSelected ? theme.colors.gray.gray300 : "transparent"};
-`;
 
 export default function RankingFilterTabs() {
   const theme = useTheme();
