+++ src/components/home/RankingSection/components/RankingHeader/index.tsx
@@ -1,0 +1,5 @@
+import { headerStyle } from "./styles";
+
+export default function RankingHeader() {
+  return <div css={headerStyle}>실시간 급상승 선물랭킹</div>;
+}
