+++ src/components/home/RankingSection/components/RankingHeader/styles.ts
@@ -1,0 +1,7 @@
+import { css } from "@emotion/react";
+
+export const headerStyle = css`
+  font-size: 1.25rem;
+  font-weight: 700;
+  padding: 16px;
+`;
