+++ src/components/home/RankingSection/components/RankingSubFilterTabs/index.tsx
@@ -1,7 +1,7 @@
-import type { Theme } from "@emotion/react";
-import { css, useTheme } from "@emotion/react";
+import { useTheme } from "@emotion/react";
 import { useSearchParams } from "react-router-dom";
 import { useEffect } from "react";
+import { containerStyle, tabItemStyle } from "./styles";
 
 const FILTER_OPTIONS = ["받고 싶어한", "많이 선물한", "위시로 받은"];
 
@@ -42,20 +42,3 @@ export default function RankingSubFilterTabs() {
     </div>
   );
 }
-
-const containerStyle = (theme: Theme) => css`
-  display: flex;
-  justify-content: space-around;
-  background-color: ${theme.colors.blue.blue100};
-  border-radius: 12px;
-  padding: 16px 0;
-`;
-
-const tabItemStyle = (theme: Theme, isActive: boolean) => css`
-  background: none;
-  border: none;
-  font-size: ${theme.typography.body1Bold.fontSize};
-  font-weight: ${isActive ? theme.typography.body1Bold.fontWeight : 400};
-  color: ${isActive ? theme.colors.blue.blue700 : theme.colors.gray.gray400};
-  cursor: pointer;
-`;
