+++ src/components/home/RankingSection/styles.ts
@@ -1,0 +1,5 @@
+import { css } from "@emotion/react";
+
+export const sectionStyle = css`
+  margin-top: 32px;
+`;
