+++ src/components/layout/Header/styles.ts
@@ -1,0 +1,22 @@
+import { css } from "@emotion/react";
+
+export const headerWrapper = () => css`
+  display: flex;
+  align-items: center;
+  padding: 12px 16px;
+  background-color: #f5f5f5;
+`;
+
+export const titleStyle = () => css`
+  flex: 1;
+  font-size: 20px;
+  font-weight: bold;
+  text-align: center;
+`;
+
+export const backButtonStyle = css`
+  cursor: pointer;
+  background: none;
+  border: none;
+  font-size: 16px;
+`;
